Fix elflist's alignment problems
This time for sure.
I'm not sure if it is a bug in the GNULD linker or if it's my fault.
In any case, I assumed that if a linker would encounter a section
multiple times, it will combine it into one. No problems here, this was
correct. But, take a look at this:

 .elflist_consoles_begin
                0x00000000c01110fc        0x4 /home/maximl/Development/my/yaeos/build/i686/kernel/arch/i686/kernel.c.o
                0x00000000c01110fc                __elflist_consoles_begin
 .elflist_consoles_begin
                0x00000000c0111100        0x4 /home/maximl/Development/my/yaeos/build/i686/kernel/kernel/console.c.o
 .elflist_consoles_begin
                0x00000000c0111104        0x4 /home/maximl/Development/my/yaeos/build/i686/kernel/modules/vga.c.o
 .elflist_consoles_data
                0x00000000c0111108        0x4 /home/maximl/Development/my/yaeos/build/i686/kernel/arch/i686/kernel.c.o
                0x00000000c0111108                __elflist_consoles_c

In this situation, the linker will place the .elflist_consoles_begin
section at the 0xc01110fc address, but next section begins at 0xc0111108!
This is not something I would expect to happen. And thus this was the
problem.

Circumvented it by adding separate macros for declaration of the
section, and getting rid of weak symbols.